added logic to map
