+++ src/main.rs
@@ -1,3 +1,4 @@
+use ggez::graphics::Color;
 use ggez::ContextBuilder;
 use ggez::{conf, event, graphics, timer, Context, GameResult};
 use lazy_static::lazy_static;
@@ -7,6 +8,9 @@ use rand::Rng;
 use std::cmp::Ordering;
 use std::collections::{BinaryHeap, HashMap};
 use std::time::Duration;
+
+mod map;
+use crate::map::Map;
 
 const MAP_SIZE: usize = 10;
 const CELL_SIZE: f32 = 30.0;
@@ -27,17 +31,16 @@ const DIRECTIONS: [(isize, isize); 8] = [
     (1, 1),
 ];
 
-lazy_static! {
-    static ref DEFAULT_COLOR: graphics::Color = graphics::Color::from_rgb(255, 255, 255);
-    static ref OBSTACLE_COLOR: graphics::Color = graphics::Color::from_rgb(127, 127, 127);
-    static ref ENERGY_COLOR: graphics::Color = graphics::Color::from_rgb(255, 255, 0);
-    static ref MINERALS_COLOR: graphics::Color = graphics::Color::from_rgb(0, 0, 255);
-    static ref SCIENCE_INTERESTS_COLOR: graphics::Color = graphics::Color::from_rgb(0, 255, 0);
-    static ref ROBOT_EXPLORER_COLOR: graphics::Color = graphics::Color::from_rgb(255, 0, 0);
-    static ref ROBOT_EXTRACTOR_COLOR: graphics::Color = graphics::Color::from_rgb(255, 165, 0);
-    static ref STATION_COLOR: graphics::Color = graphics::Color::from_rgb(0, 255, 255);
-    static ref FOG_COLOR: graphics::Color = graphics::Color::from_rgb(0, 0, 0);
-}
+// Constantes de couleurs
+const DEFAULT_COLOR: Color = Color::new(1.0, 1.0, 1.0, 1.0); // White
+const OBSTACLE_COLOR: Color = Color::new(0.498, 0.498, 0.498, 1.0); // Gray
+const ENERGY_COLOR: Color = Color::new(1.0, 1.0, 0.0, 1.0); // Yellow
+const MINERALS_COLOR: Color = Color::new(0.0, 0.0, 1.0, 1.0); // Blue
+const SCIENCE_INTERESTS_COLOR: Color = Color::new(0.0, 1.0, 0.0, 1.0); // Green
+const ROBOT_EXPLORER_COLOR: Color = Color::new(1.0, 0.0, 0.0, 1.0); // Red
+const ROBOT_EXTRACTOR_COLOR: Color = Color::new(1.0, 0.647, 0.0, 1.0); // Orange
+const STATION_COLOR: Color = Color::new(0.0, 1.0, 1.0, 1.0); // Cyan
+const FOG_COLOR: Color = Color::new(0.0, 0.0, 0.0, 1.0); // Black
 
 #[derive(Copy, Clone, Eq, PartialEq)]
 struct Node {
@@ -68,21 +71,6 @@ impl PartialOrd for Node {
     fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
         Some(self.cmp(other))
     }
-}
-
-struct Map {
-    energy: [[bool; MAP_SIZE]; MAP_SIZE],
-    minerals: [[bool; MAP_SIZE]; MAP_SIZE],
-    science_interests: [[bool; MAP_SIZE]; MAP_SIZE],
-    obstacles: [[bool; MAP_SIZE]; MAP_SIZE],
-    explored: [[bool; MAP_SIZE]; MAP_SIZE],
-    fog_of_war: [[bool; MAP_SIZE]; MAP_SIZE],
-    robot_explorer: Option<RobotExplorer>,
-    robot_extractor: Option<RobotExtractor>,
-    update_timer: Duration,
-    game_over: bool,
-    nb_consumables: usize,
-    resources: Vec<(usize, usize)>,
 }
 
 struct RobotExplorer {
@@ -235,21 +223,11 @@ impl RobotExplorer {
             if map.energy[new_y][new_x] {
                 self.founded_resource = true;
                 self.resource_position = Some((new_x, new_y)); // Ajouter cette ligne
-                println!(
-                    "Founded energy at ({}, {}). Remaining: {}",
-                    new_x,
-                    new_y,
-                    map.count_consumables()
-                );
+                println!("Founded energy at ({}, {}).", new_x, new_y,);
             } else if map.minerals[new_y][new_x] {
                 self.founded_resource = true;
                 self.resource_position = Some((new_x, new_y)); // Ajouter cette ligne
-                println!(
-                    "Founded minerals at ({}, {}). Remaining: {}",
-                    new_x,
-                    new_y,
-                    map.count_consumables()
-                );
+                println!("Founded minerals at ({}, {}).", new_x, new_y,);
             }
         }
     }
@@ -312,311 +290,8 @@ impl RobotExplorer {
     }
 }
 
-impl Map {
-    fn new() -> Self {
-        let mut map = Map {
-            energy: [[false; MAP_SIZE]; MAP_SIZE],
-            minerals: [[false; MAP_SIZE]; MAP_SIZE],
-            science_interests: [[false; MAP_SIZE]; MAP_SIZE],
-            obstacles: [[false; MAP_SIZE]; MAP_SIZE],
-            explored: [[false; MAP_SIZE]; MAP_SIZE],
-            fog_of_war: [[true; MAP_SIZE]; MAP_SIZE],
-            robot_explorer: None, // Initialisez le robot à None pour l'instant
-            robot_extractor: None,
-            update_timer: Duration::from_secs(1),
-            game_over: false,
-            nb_consumables: 0,
-            resources: vec![], // Initialisez le champ des ressources
-        };
-
-        // Initialisez la position du robot en appelant la fonction init_robot_position
-        if let Some((x, y)) = map.init_robot_position() {
-            map.robot_explorer = Some(RobotExplorer::new(x, y));
-            map.robot_extractor = Some(RobotExtractor::new(x, y));
-            map.fog_of_war[y][x] = false;
-        }
-
-        map
-    }
-
-    fn init_robot_position(&self) -> Option<(usize, usize)> {
-        let mut rng = rand::thread_rng();
-        let mut attempts = 0;
-        const MAX_ATTEMPTS: usize = 100; // Nombre maximal de tentatives
-
-        loop {
-            let x = rng.gen_range(0..MAP_SIZE);
-            let y = rng.gen_range(0..MAP_SIZE);
-
-            if !self.obstacles[y][x] && !self.energy[y][x] && !self.minerals[y][x] {
-                return Some((x, y));
-            }
-
-            attempts += 1;
-            if attempts >= MAX_ATTEMPTS {
-                return None;
-            }
-        }
-    }
-
-    fn place_obstacles(&mut self, _resources: &[(usize, usize)]) {
-        let mut rng = rand::thread_rng();
-        let seed = rng.gen();
-        let fbm_obstacles = Fbm::<Perlin>::new(seed);
-
-        // Utilisez les emplacements des ressources pour éviter de placer des obstacles à des emplacements similaires
-        for y in 0..MAP_SIZE {
-            for x in 0..MAP_SIZE {
-                // Vérifiez si l'emplacement actuel est déjà occupé par une ressource
-                if !self
-                    .resources
-                    .iter()
-                    .any(|&(res_x, res_y)| res_x == x && res_y == y)
-                {
-                    let obstacles_noise = fbm_obstacles.get([x as f64, y as f64]);
-                    self.obstacles[y][x] = obstacles_noise > OBSTACLE_THRESHOLD;
-                }
-            }
-        }
-    }
-
-    fn place_resources(&mut self) {
-        let mut rng = rand::thread_rng();
-        let energy_seed = rng.gen();
-        let minerals_seed = rng.gen();
-        let fbm_energy = Fbm::<Perlin>::new(energy_seed);
-        let fbm_minerals = Fbm::<Perlin>::new(minerals_seed);
-
-        for y in 0..MAP_SIZE {
-            for x in 0..MAP_SIZE {
-                let energy_noise = fbm_energy.get([x as f64, y as f64]);
-                let minerals_noise = fbm_minerals.get([x as f64, y as f64]);
-                self.energy[y][x] = energy_noise > ENERGY_THRESHOLD;
-                self.minerals[y][x] = minerals_noise > MINERALS_THRESHOLD;
-
-                if self.energy[y][x] || self.minerals[y][x] {
-                    self.resources.push((x, y));
-                }
-
-                println!("Number of consumables: {}", self.resources.len());
-            }
-        }
-    }
-
-    fn update_robot(&mut self) {
-        if self.game_over {
-            return;
-        }
-
-        if let Some(mut extractor) = self.robot_extractor.take() {
-            if !extractor.waiting {
-                if extractor.carrying_resource {
-                    if let Some(path) = extractor.move_towards(
-                        extractor.station_x,
-                        extractor.station_y,
-                        &self.obstacles,
-                    ) {
-                        if path.len() > 1 {
-                            let (next_x, next_y) = path[1];
-                            extractor.x = next_x;
-                            extractor.y = next_y;
-
-                            if extractor.x == extractor.station_x
-                                && extractor.y == extractor.station_y
-                            {
-                                extractor.carrying_resource = false;
-                                extractor.waiting = true; // Extractor attend maintenant
-                                if let Some(mut explorer) = self.robot_explorer.as_mut() {
-                                    explorer.waiting = false; // L'explorer peut maintenant bouger
-                                }
-                                println!("Extractor returned to the station with resource.");
-                            }
-                        }
-                    }
-                } else if let Some((target_x, target_y)) = extractor.target_position {
-                    if let Some(path) = extractor.move_towards(target_x, target_y, &self.obstacles)
-                    {
-                        if path.len() > 1 {
-                            let (next_x, next_y) = path[1];
-                            extractor.x = next_x;
-                            extractor.y = next_y;
-
-                            if extractor.x == target_x && extractor.y == target_y {
-                                extractor.collect_resource();
-                                self.energy[target_y][target_x] = false;
-                                self.minerals[target_y][target_x] = false;
-                            }
-                        }
-                    }
-                }
-            }
-            self.robot_extractor = Some(extractor);
-        }
-
-        if let Some(mut explorer) = self.robot_explorer.take() {
-            if !explorer.waiting {
-                let mut rng = rand::thread_rng();
-
-                let all_resources_collected = self.count_consumables() == 0 || self.is_map_empty();
-
-                if all_resources_collected {
-                    if let Some(path) = explorer.return_to_station(&self.obstacles) {
-                        if path.len() > 1 {
-                            let (next_x, next_y) = path[1];
-                            explorer.x = next_x;
-                            explorer.y = next_y;
-
-                            if explorer.x == explorer.station_x && explorer.y == explorer.station_y
-                            {
-                                self.game_over = true;
-                                println!("Game Over: Robot returned to the station. All resources have been collected.");
-                            }
-                        }
-                    }
-                } else if explorer.founded_resource {
-                    if let Some((resource_x, resource_y)) = explorer.resource_position {
-                        if explorer.x == explorer.station_x && explorer.y == explorer.station_y {
-                            self.robot_extractor.as_mut().unwrap().target_position =
-                                Some((resource_x, resource_y));
-                            explorer.founded_resource = false;
-                            explorer.resource_position = None;
-                            explorer.waiting = true; // Explorer attend maintenant
-                            if let Some(mut extractor) = self.robot_extractor.as_mut() {
-                                extractor.waiting = false; // L'extractor peut maintenant bouger
-                            }
-                            println!("Explorer returned to the station and provided resource position to Extractor.");
-                        } else if let Some(path) = explorer.return_to_station(&self.obstacles) {
-                            if path.len() > 1 {
-                                let (next_x, next_y) = path[1];
-                                explorer.x = next_x;
-                                explorer.y = next_y;
-                            }
-                        }
-                    }
-                } else {
-                    explorer.move_random(&mut rng, self);
-                }
-            }
-            self.robot_explorer = Some(explorer);
-        }
-    }
-
-    fn is_map_empty(&self) -> bool {
-        !self.energy.iter().any(|row| row.iter().any(|&val| val))
-            && !self.minerals.iter().any(|row| row.iter().any(|&val| val))
-    }
-
-    fn count_consumables(&self) -> usize {
-        self.energy
-            .iter()
-            .flatten()
-            .chain(self.minerals.iter().flatten())
-            .filter(|&&val| val)
-            .count()
-    }
-}
-
 fn heuristic(x1: usize, y1: usize, x2: usize, y2: usize) -> usize {
     ((x2 as isize - x1 as isize).abs() + (y1 as isize - y2 as isize).abs()) as usize
-}
-
-impl event::EventHandler for Map {
-    fn draw(&mut self, ctx: &mut Context) -> GameResult {
-        graphics::clear(ctx, *DEFAULT_COLOR);
-
-        for y in 0..MAP_SIZE {
-            for x in 0..MAP_SIZE {
-                let mut color = if self.fog_of_war[y][x] {
-                    *FOG_COLOR
-                } else {
-                    *DEFAULT_COLOR
-                };
-
-                if !self.fog_of_war[y][x] {
-                    match (
-                        self.energy[y][x],
-                        self.minerals[y][x],
-                        self.science_interests[y][x],
-                        self.obstacles[y][x],
-                    ) {
-                        (true, false, false, false) => color = *ENERGY_COLOR,
-                        (false, true, false, false) => color = *MINERALS_COLOR,
-                        (false, false, true, false) => color = *SCIENCE_INTERESTS_COLOR,
-                        (false, false, false, true) => color = *OBSTACLE_COLOR,
-                        _ => color = *DEFAULT_COLOR,
-                    }
-                }
-
-                let rect = graphics::Rect::new(
-                    x as f32 * CELL_SIZE,
-                    y as f32 * CELL_SIZE,
-                    CELL_SIZE,
-                    CELL_SIZE,
-                );
-                let mesh =
-                    graphics::Mesh::new_rectangle(ctx, graphics::DrawMode::fill(), rect, color)?;
-                graphics::draw(ctx, &mesh, graphics::DrawParam::default())?;
-            }
-        }
-
-        if let Some(robot) = &self.robot_explorer {
-            let station_rect = graphics::Rect::new(
-                robot.station_x as f32 * CELL_SIZE,
-                robot.station_y as f32 * CELL_SIZE,
-                CELL_SIZE,
-                CELL_SIZE,
-            );
-            let station_mesh = graphics::Mesh::new_rectangle(
-                ctx,
-                graphics::DrawMode::fill(),
-                station_rect,
-                *STATION_COLOR,
-            )?;
-            graphics::draw(ctx, &station_mesh, graphics::DrawParam::default())?;
-        }
-
-        if let Some(robot) = &self.robot_explorer {
-            let robot_rect = graphics::Rect::new(
-                robot.x as f32 * CELL_SIZE,
-                robot.y as f32 * CELL_SIZE,
-                CELL_SIZE,
-                CELL_SIZE,
-            );
-            let robot_mesh = graphics::Mesh::new_rectangle(
-                ctx,
-                graphics::DrawMode::fill(),
-                robot_rect,
-                *ROBOT_EXPLORER_COLOR,
-            )?;
-            graphics::draw(ctx, &robot_mesh, graphics::DrawParam::default())?;
-        }
-
-        if let Some(robot) = &self.robot_extractor {
-            let robot_rect = graphics::Rect::new(
-                robot.x as f32 * CELL_SIZE,
-                robot.y as f32 * CELL_SIZE,
-                CELL_SIZE,
-                CELL_SIZE,
-            );
-            let robot_mesh = graphics::Mesh::new_rectangle(
-                ctx,
-                graphics::DrawMode::fill(),
-                robot_rect,
-                *ROBOT_EXTRACTOR_COLOR,
-            )?;
-            graphics::draw(ctx, &robot_mesh, graphics::DrawParam::default())?;
-        }
-
-        graphics::present(ctx)?;
-        Ok(())
-    }
-
-    fn update(&mut self, ctx: &mut Context) -> GameResult {
-        if timer::check_update_time(ctx, 1) {
-            self.update_robot();
-        }
-        Ok(())
-    }
 }
 
 fn main() -> GameResult {
